Move source files into subdirectory, update setup configuration
