+++ pyproject.toml
@@ -1,0 +1,8 @@
+[build-system]
+requires = [
+    "setuptools>=42",
+    "wheel",
+    "setuptools_scm",
+    "pytest>=5,<8",
+]
+build-backend = "setuptools.build_meta"
