+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 asyncio_mode=strict
+addopts = --doctest-modules
