+++ setup.cfg
@@ -1,5 +1,27 @@
 [metadata]
-description-file = README.md
+name = volkswagencarnet
+version = 1.0.0
+author = Robin Ostlund
+author_email = me@robinostlund.name
+description = Communicate with Volkswagen WeConnect
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/robinostlund/volkswagencarnet
+project_urls =
+    Bug Tracker = https://github.com/robinostlund/volkswagencarnet/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: Apache License 2.0
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+  = .
+packages = find:
+python_requires = >= 3.6
+
+[options.packages.find]
+where = .
 
 [flake8]
 ignore =
