--- setup.py
@@ -1,32 +1,0 @@
-# read the contents of your README file
-from os import path
-from setuptools import setup
-
-this_directory = path.abspath(path.dirname(__file__))
-with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
-    long_description = f.read()
-
-setup(
-    name='volkswagencarnet',
-    description='Communicate with Volkswagen WeConnect',
-    author='Robin Ostlund',
-    author_email='me@robinostlund.name',
-    url='https://github.com/robinostlund/volkswagencarnet',
-    long_description=long_description,
-    long_description_content_type='text/markdown',
-    py_modules=[
-        "vw_connection",
-        "vw_vehicle",
-        "vw_dashboard",
-        "vw_utilities",
-        "vw_const",
-        "__init__"
-    ],
-    provides=["volkswagencarnet"],
-    install_requires=list(open("requirements.txt").read().strip().split("\n")),
-    use_scm_version=True,
-    setup_requires=[
-        'setuptools_scm',
-        'pytest>=5,<8',
-    ]
-)
