+++ tests/dummy_test.py
@@ -1,17 +1,12 @@
 import pytest
-import sys
-import os
 from aiohttp import ClientSession
 
 # we need to change os path to be able to import volkswagecarnet
-myPath = os.path.dirname(os.path.abspath(__file__))
-print(myPath)
-sys.path.insert(0, myPath + '/../')
+from volkswagencarnet import vw_connection
 
 
 @pytest.mark.asyncio
 async def test_volkswagencarnet():
-    import vw_connection
     async with ClientSession() as session:
         connection = vw_connection.Connection(session, 'test@example.com', 'test_password')
         # if await connection._login():
