+++ volkswagencarnet/vw_connection.py
@@ -17,13 +17,14 @@ from json import dumps as to_json
 import aiohttp
 from bs4 import BeautifulSoup
 from base64 import b64encode
-from vw_utilities import read_config, json_loads
-from vw_vehicle import Vehicle
 
 from aiohttp import ClientSession, ClientTimeout
 from aiohttp.hdrs import METH_GET, METH_POST
 
-from vw_const import (
+from .vw_utilities import json_loads, read_config
+from .vw_vehicle import Vehicle
+
+from .vw_const import (
     BRAND,
     COUNTRY,
     HEADERS_SESSION,
